Updating error handler to use the appropriate regex
-- What --
It appears that nearlyError2Object was expecting a different output
for errors and was attempting to match based on the wrong regex (there
appear to be two similarly named regex patterns for parsing error
messages). In cases where the regex failed, the program still
attempted to extract information from it, even though in such cases
the result would be undefined.

This commit switched to using the alternate available error regex and
satisfies the example posted in Issue #6.